MachineLearning&DataMining/Labs/lab3/task/example-resnet-34/ResNet34-code/requirements.txt
matplotlib==3.4.3
Pillow==9.1.0
tqdm==4.62.3
